requirements.txt
flet>=0.21.0
groq>=0.4.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
